+++ fibonacci/src/main/java/com/test/fibonacci/infrastructure/entrypoints/dtos/NthNumberResponseDto.java
@@ -1,6 +1,6 @@
 package com.test.fibonacci.infrastructure.entrypoints.dtos;
 
 public record NthNumberResponseDto(
-        Integer nthNumber
+        Long nthNumber
 ) {
 }
